lab/pyproject.toml
[tool.poetry]
name = "lab"
version = "0.1.0"
description = ""
authors = ["Doarakko"]

[tool.poetry.dependencies]
python = "^3.7"
elasticsearch = "^7.9.1"
bert-serving-server = "^1.10.0"
bert-serving-client = "^1.10.0"
sentencepiece = "^0.1.92"
pandas = "^1.1.3"
ginza = "^4.0.5"
jupyterlab = "^2.2.8"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
